+++ src/components/index.ts
@@ -6,3 +6,4 @@ export * from './UserList';
 export * from './PanelList';
 export * from './PanelItem';
 export * from './PanelValriables';
+export * from './CreateScheduleFormPartial';
